Renamed : the feeds folder to posts folder, initialized the 'create new post' page
